Renamed Examples to example, updated namespace for the same
